+++ src/onegov.election_day/onegov/election_day/tests/layouts/test_vote_layout.py
@@ -54,6 +54,8 @@ def test_vote_layout(session):
     assert layout.type == 'simple'
     assert layout.main_view == 'Vote/entities'
     assert layout.ballot.type == 'proposal'
+    assert layout.entities_map_link == 'Vote/proposal-by-entities-map'
+    assert layout.districts_map_link == 'Vote/proposal-by-districts-map'
 
     layout = VoteLayout(
         ComplexVote(), DummyRequest(), tab='counter-proposal-entities'
@@ -61,6 +63,10 @@ def test_vote_layout(session):
     assert layout.type == 'complex'
     assert layout.main_view == 'ComplexVote/proposal-entities'
     assert layout.ballot.type == 'counter-proposal'
+    assert layout.entities_map_link == \
+        'ComplexVote/counter-proposal-by-entities-map'
+    assert layout.districts_map_link == \
+        'ComplexVote/counter-proposal-by-districts-map'
 
     with freeze_time("2014-01-01 12:00"):
         vote = ComplexVote(